[#17] Rename Drawer to MainScreen, extract MainScreen out of MainActivity
