+++ android/src/main/java/vn/luongvo/kmm/survey/android/ui/screens/MainScreen.kt
@@ -1,12 +1,11 @@
-package vn.luongvo.kmm.survey.android.ui.screens.home.views
+package vn.luongvo.kmm.survey.android.ui.screens
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.shape.CircleShape
-import androidx.compose.material.Divider
-import androidx.compose.material.Text
-import androidx.compose.runtime.Composable
+import androidx.compose.material.*
+import androidx.compose.runtime.*
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.Color.Companion.White
@@ -16,8 +15,12 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import coil.compose.AsyncImage
+import kotlinx.coroutines.launch
+import timber.log.Timber
 import vn.luongvo.kmm.survey.android.BuildConfig
 import vn.luongvo.kmm.survey.android.R
+import vn.luongvo.kmm.survey.android.ui.common.RtlModalDrawer
+import vn.luongvo.kmm.survey.android.ui.navigation.AppNavigation
 import vn.luongvo.kmm.survey.android.ui.screens.home.HomeUserAvatar
 import vn.luongvo.kmm.survey.android.ui.screens.home.UserUiModel
 import vn.luongvo.kmm.survey.android.ui.theme.*
@@ -25,7 +28,40 @@ import vn.luongvo.kmm.survey.android.ui.theme.AppTheme.dimensions
 import vn.luongvo.kmm.survey.android.ui.theme.AppTheme.typography
 
 @Composable
-fun Drawer(
+fun MainScreen(
+    initialDrawerValue: DrawerValue = DrawerValue.Closed
+) {
+    val drawerState = rememberDrawerState(initialDrawerValue)
+    val scope = rememberCoroutineScope()
+    val openDrawer = {
+        scope.launch {
+            drawerState.open()
+        }
+    }
+    var user by remember { mutableStateOf<UserUiModel?>(null) }
+
+    RtlModalDrawer(
+        drawerState = drawerState,
+        gesturesEnabled = drawerState.isOpen,
+        drawerContent = {
+            Drawer(
+                user = user,
+                onLogoutClick = {
+                    // TODO https://github.com/luongvo/kmm-survey/issues/19
+                    Timber.d("onLogoutClick")
+                }
+            )
+        }
+    ) {
+        AppNavigation(
+            onDrawerUiStateChange = { user = it },
+            onOpenDrawer = { openDrawer() }
+        )
+    }
+}
+
+@Composable
+private fun Drawer(
     user: UserUiModel?,
     onLogoutClick: () -> Unit,
     modifier: Modifier = Modifier
@@ -75,6 +111,16 @@ fun Drawer(
             text = "v${BuildConfig.VERSION_NAME} (${BuildConfig.VERSION_CODE})",
             color = White50,
             style = typography.subtitle1.copy(fontSize = 11.sp)
+        )
+    }
+}
+
+@Preview
+@Composable
+fun MainScreenPreview() {
+    ComposeTheme {
+        MainScreen(
+            initialDrawerValue = DrawerValue.Open
         )
     }
 }
